pyproject.toml
[tool.poetry]
name = "party-up"
version = "0.1.0"
description = ""
authors = ["bbakersmith <b@baker-smith.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyserial = "^3.5"
urwid = "^2.4.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
black = "^23.12.1"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
